docker/backend/requirements.txt
aerich==0.8.0
asyncpg==0.30.0
passlib==1.7.4
fastapi[standard]==0.115.6
tortoise-orm==0.22.2
uvicorn==0.34.0
pydantic-settings==2.7.0
python-jose==3.3.0
python-multipart==0.0.20